pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lite-kits"
version = "0.1.1"
description = "Lightweight enhancement kits for vanilla dev tools (spec-kit, etc.)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "lite-kits contributors" }
]
keywords = ["spec-kit", "enhancement", "kits", "ai", "development-workflow", "git"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
lite-kits = "lite_kits.cli:app"

[project.urls]
Homepage = "https://github.com/tmorgan181/lite-kits"
Documentation = "https://github.com/tmorgan181/lite-kits#readme"
Repository = "https://github.com/tmorgan181/lite-kits"
Issues = "https://github.com/tmorgan181/lite-kits/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lite_kits"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
